refac: mostly do away with the Architecture trait
